improve key input handling
